requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
sqlitecloud==0.0.78
starlette==0.37.2
tinytuya==1.15.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
numpy==2.0.1
passlib==1.7.4
bcrypt==4.0.1